requirements.txt
timm==0.6.12
opencv-python==4.5.5.64
tqdm==4.50.2
torchstat==0.0.7
torchsummary==1.5.1
Pillow==8.3.1
h5py==2.10.0
thop==0.1.1.post2209072238
imageio==2.9.0


